# readers/openapi.yml
openapi: 3.0.1
info:
  title: Mainflux reader service
  description: HTTP API for reading messages.
  version: "1.0.0"

paths:
  /channels/{chanId}/messages:
    get:
      summary: Retrieves messages sent to single channel
      description: |
        Retrieves a list of messages sent to specific channel. Due to
        performance concerns, data is retrieved in subsets. The API readers must
        ensure that the entire dataset is consumed either by making subsequent
        requests, or by increasing the subset size of the initial request.
      tags:
        - messages
      parameters:
        - $ref: "#/components/parameters/Authorization"
        - $ref: "#/components/parameters/ChanId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Publisher"
        - $ref: "#/components/parameters/Name"
        - $ref: "#/components/parameters/Value"
        - $ref: "#/components/parameters/BoolValue"
        - $ref: "#/components/parameters/StringValue"
        - $ref: "#/components/parameters/DataValue"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
      responses:
        '200':
          $ref: "#/components/responses/MessagesPageRes"
        '400':
          description: Failed due to malformed query parameters.
        '403':
          description: Missing or invalid access token provided.
        '500':
          $ref: "#/components/responses/ServiceError"
  /messages/list/{chanID}:
    get:
      summary: Retrieves timeseries messages sent by things
      description: |
        Retrieves a list of timeseries messages. Due to performance concerns,
        data is retrieved in subsets. The API readers must ensure that the
        entire dataset is consumed either by making subsequent requests, or
        by increasing the subset size of the initial request.
      tags:
        - messages
      operationId: findMessagesByPublisher
      parameters:
        - $ref: "#/components/parameters/ChanId"
        - $ref: "#/components/parameters/Name"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - $ref: "#/parameters/AggregationType"
        - $ref: "#/parameters/Interval"
      responses:
        200:
          #description: Data retrieved.
          $ref: "#/components/responses/MessagesPageRes"
        400:
          description: Failed due to malformed query parameters.
        403:
          description: Missing or invalid access token provided.
        500:
          $ref: "#/responses/ServiceError"
  /messages/last/{chanIDs}:
    get:
      summary: Retrieves last timeseries messages sent by things identified by a list of publishers.
      description: |
        Retrieves last timeseries messages sent by things identified by a list of publishers.
        Returned timeseries data can be grouped by sensor name or subtopic.
      tags:
        - messages
      operationId: getLastMeasurement
      parameters:
        - $ref: "#/components/parameters/ChanIds"
        - $ref: "#/parameters/Name"
      responses:
        200:
          #description: Data retrieved.
          $ref: "#/components/responses/MessagesPageRes"
        400:
          description: Failed due to malformed query parameters.
        403:
          description: Missing or invalid access token provided.
        500:
          $ref: "#/responses/ServiceError"

  /messages/pumpRunningSeconds/{chanIDs}:
    post:
      summary: 获取指定泵站的开启时间
      description: |
        统计多个指定泵站 每个泵站在给定时间(startTime, endTime)内开启的总时间
        返回结果:TimeseriesData
      tags:
        - messages
      operationId: pumpRunningSeconds
      parameters:
        - $ref: "#/components/parameters/ChanIds"
        - $ref: "#/parameters/Names"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
      responses:
        200:
          #description: Data retrieved.
          $ref: "#/components/responses/MessagesPageRes"
        400:
          description: Failed due to malformed query parameters.
        403:
          description: Missing or invalid access token provided.
        500:
          $ref: "#/responses/ServiceError"

components:
  schemas:
    MessagesPage:
      type: object
      properties:
        total:
          type: number
          description: Total number of items that are present on the system.
        offset:
          type: number
          description: Number of items that were skipped during retrieval.
        limit:
          type: number
          description: Size of the subset that was retrieved.
        messages:
          type: array
          minItems: 0
          uniqueItems: true
          items:
            type: object
            properties:
              channel:
                type: integer
                description: Unique channel id.
              publisher:
                type: integer
                description: Unique publisher id.
              protocol:
                type: string
                description: Protocol name.
              name:
                type: string
                description: Measured parameter name.
              unit:
                type: string
                description: Value unit.
              value:
                type: number
                description: Measured value in number.
              stringValue:
                type: string
                description: Measured value in string format.
              boolValue:
                type: boolean
                description: Measured value in boolean format.
              dataValue:
                type: string
                description: Measured value in binary format.
              valueSum:
                type: number
                description: Sum value.
              time:
                type: number
                description: Time of measurement.
              updateTime:
                type: number
                description: Time of updating measurement.
    TimeseriesData:
      type: object
      properties:
        messages:
          type: array
          minItems: 0
          uniqueItems: true
          items:
            $ref: "#/definitions/TimeseriesDataArray"
    TimeseriesDataArray:
      type: object
      properties:
        publisher:
          type: string
          description: Publisher ID.
        name:
          type: string
          description: Name of the sensor.
        subtopic:
          type: string
          description: Message subtopic.
        unit:
          type: string
          description: Unit of the measurement.
        total:
          type: integer
          description: Total number of datapoints.
        data:
          type: array
          minItems: 0
          items:
            $ref: "#/definitions/TimeseriesDataPoint"
    TimeseriesDataPoint:
      type: object
      properties:
        value:
          type: number
          description: Measured value in number.
        time:
          type: string
          description: Time of measurement.
  parameters:
    Authorization:
      name: Authorization
      description: Thing access token.
      in: header
      schema:
        type: string
      required: true
    ChanId:
      name: chanId
      description: Unique channel identifier.
      in: path
      schema:
        type: string
        format: uuid
      required: true
    Limit:
      name: limit
      description: Size of the subset to retrieve.
      in: query
      schema:
        type: integer
        default: 10
        maximum: 100
        minimum: 1
      required: false
    Offset:
      name: offset
      description: Number of items to skip during retrieval.
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
      required: false
    Publisher:
      name: Publisher
      description: Unique thing identifier.
      in: query
      schema:
        type: string
        format: uuid
      required: false
    Name:
      name: name
      description: SenML message name.
      in: query
      schema:
        type: string
      required: false
    Value:
      name: v
      description: SenML message value.
      in: query
      schema:
        type: string
      required: false
    BoolValue:
      name: vb
      description: SenML message bool value.
      in: query
      schema:
        type: boolean
      required: false
    StringValue:
      name: vs
      description: SenML message string value.
      in: query
      schema:
        type: string
      required: false
    DataValue:
      name: vd
      description: SenML message data value.
      in: query
      schema:
        type: string
      required: false
    From:
      name: from
      description: SenML message time in nanoseconds (integer part represents seconds).
      in: query
      schema:
        type: float
      required: false
    To:
      name: to
      description: SenML message time in nanoseconds (integer part represents seconds).
      in: query
      schema:
        type: float
      required: false
    Publishers:
      name: publishers
      description: A comma-separated list of publisher identifiers.
      in: path
      schema:
        type: string
      required: true
    SensorName:
      name: sensorName
      description: Filter by name of the sensor.
      in: query
      schema:
        type: string
    SensorNames:
      name: SensorNames
      description: SensorName list
      in: query
      schema:
        type: string
    SubTopic:
      name: subTopic
      description: Filter by name of the subtopic.
      in: query
      schema:
        type: string
    StartTime:
      name: startTime
      description: The start timestamp of the requested timeframe, in UTC milliseconds.
      in: query
      schema:
        type: string
    EndTime:
      name: endTime
      description: The end timestamp of the requested timeframe, in UTC milliseconds.
      in: query
      schema:
        type: string
    AggregationType:
      name: aggregationType
      description: The aggregation type for the resulting data points.
      in: query
      schema:
        type: string
#      enum:
#        - AVG
#        - COUNT
#        - MAX
#        - MEDIAN
#        - MIN
#        - PERCENTILE
#        - SUM
    Interval:
      name: interval
      description: Time interval
      in: query
      schema:
        type: string

  responses:

    MessagesPageRes:
      description: Data retrieved.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MessagesPage"

    ServiceError:
      description: Unexpected server-side error occurred.